pyproject.toml
[tool.poetry]
name = "ara-flask"
version = "0.1.0"
description = ""
authors = ["Kyle Zheng <kylezheng73@gmail.com>"]
readme = "README.md"
packages = [{include = "ara_flask"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
sqlalchemy = "^1.4.46"
sqlalchemy-cockroachdb = "^1.4.4"
cockroachdb = "^0.3.5"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
